server/pyproject.toml
[tool.poetry]
name = "grpc-api"
version = "0.1.0"
description = ""
authors = ["kkaarroollm <mkarol.4514@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.110.2"
pydantic = "^2.7.0"
grpcio = "^1.62.2"
grpcio-tools = "^1.62.2"
pydantic-settings = "^2.2.1"
uvicorn = "^0.29.0"
python-dotenv = "^1.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
